6차 스프린트 적용

+++ lib/resources/providers/accountProvider.dart
@@ -3,20 +3,20 @@ import 'dart:convert';
 
 import 'package:meta/meta.dart';
 import 'package:http/http.dart' as http;
+import 'package:shared_preferences/shared_preferences.dart';
 
 import 'package:Eliverd/models/models.dart';
 
 class AccountAPIClient {
-  // TO-DO: 백엔드 API 공식 배포 후 수정
-  static const baseUrl = 'http://localhost:8000';
+  static const baseUrl = 'SECRET:8000';
   final http.Client httpClient;
 
   AccountAPIClient({
     @required this.httpClient,
   }) : assert(httpClient != null);
 
-  Future<User> signUpUser(Map<String, dynamic> jsonifiedUser) async {
-    final url = '$baseUrl/account/user';
+  Future<void> signUpUser(Map<String, dynamic> jsonifiedUser) async {
+    final url = '$baseUrl/account/user/';
     final res = await this.httpClient.post(
       url,
       body: jsonifiedUser,
@@ -26,19 +26,28 @@ class AccountAPIClient {
     if (res.statusCode != 201) {
       throw Exception('Error occurred while registering user');
     }
-
-    final data = json.decode(res.body) as User;
-
-    return data;
   }
 
-  Future<Session> createSession(String userId, String password) async {
+  Future<String> createSession([String userId, String password]) async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+
+    final currentSession = prefs.getString('session');
+
+    if (currentSession != null) {
+      return currentSession;
+    } else if (userId == null || password == null) {
+      return null;
+    }
+
     final Map<String, dynamic> user = {
-      'user_id': userId,
-      'password': password
+      'user_id': prefs.getString('userId') ?? userId,
+      'password': prefs.getString('password') ?? password,
     };
 
-    final url = '$baseUrl/account/session';
+    prefs.setString('userId', userId);
+    prefs.setString('password', password);
+
+    final url = '$baseUrl/account/session/';
     final res = await this.httpClient.post(
       url,
       body: user,
@@ -46,38 +55,60 @@ class AccountAPIClient {
     );
 
     if (res.statusCode != 201) {
+      prefs.remove('session');
       throw Exception('Error occurred while creating session');
     }
 
-    final session = json.decode(res.body)['session'];
+    final jsonData = utf8.decode(res.bodyBytes);
+
+    final session = json.decode(jsonData)['session'] as String;
+
+    prefs.setString('session', session);
 
     return session;
   }
 
-  Future<Map<String, dynamic>> validateSession(int session) async {
-    final url = '$baseUrl/account/session';
+  Future<Map<String, dynamic>> validateSession() async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+
+    final currentSession = prefs.getString('session');
+
+    final url = '$baseUrl/account/session/';
     final res = await this.httpClient.get(url,
         headers: {
-          'Authorization': session.toString(),
+          'Authorization': currentSession,
         }
     );
 
     if (res.statusCode != 200) {
+      prefs.remove('session');
       throw Exception('Error occurred while validating session');
     }
 
-    final userInfo = json.decode(res.body);
+    final jsonData = utf8.decode(res.bodyBytes);
+
+    final userInfo = json.decode(jsonData);
 
     return userInfo;
   }
 
-  Future<void> deleteSession(int session) async {
-    final url = '$baseUrl/account/session';
+  Future<void> deleteSession() async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+
+    final currentSession = prefs.getString('session');
+
+    if (currentSession == null) {
+      return;
+    }
+
+    final url = '$baseUrl/account/session/';
     final res = await this.httpClient.delete(url,
       headers: {
-        'Authorization': session.toString(),
+        'Authorization': currentSession,
       }
     );
+
+    await prefs.remove('session');
 
     if (res.statusCode != 204) {
       throw Exception('Error occurred while deleting session');
@@ -85,34 +116,40 @@ class AccountAPIClient {
   }
 
   Future<Map<String, dynamic>> validateUser(Map<String, dynamic> jsonifiedUser) async {
-    final url = '$baseUrl/account/user/validate';
+    final url = '$baseUrl/account/user/validate/';
+
     final res = await this.httpClient.post(
       url,
       body: jsonifiedUser,
       encoding: Encoding.getByName('application/json; charset=\'utf-8\''),
     );
 
-    if (res.statusCode != 201) {
+    if (res.statusCode != 200) {
       throw Exception('Error occurred while validating user');
     }
 
-    final data = json.decode(res.body);
+    final jsonData = utf8.decode(res.bodyBytes);
+
+    final data = json.decode(jsonData);
 
     return data;
   }
 
   Future<List<User>> searchUser(String keyword) async {
-    final url = '$baseUrl/account/user/search/$keyword?is_seller=true';
+    final url = '$baseUrl/account/user/search/$keyword/?is_seller=True';
     final res = await this.httpClient.get(url);
 
-    if (res.statusCode != 201) {
-      throw Exception('Error occurred while searching user');
+    if (res.statusCode != 200) {
+      return <User>[];
     }
 
-    final data = json.decode(res.body)['results'] as List;
+    final jsonData = utf8.decode(res.bodyBytes);
+
+    final data = json.decode(jsonData)['results'] as List;
 
     return data.map((rawUser) {
       return User(
+        pid: rawUser['pid'],
         userId: rawUser['user_id'],
         nickname: rawUser['nickname'],
         realname: rawUser['realname'],
